remove old implementations of contracts
