+++ src/starknet.rs
@@ -1,4 +1,4 @@
 #![allow(unused)]
 pub mod admin_account;
-pub mod main_contract;
+pub mod operator_contract;
 pub mod provider;
